split todo list into own file
